refactoring day11
