+++ 2019/src/lib.rs
@@ -10,3 +10,4 @@ pub mod day07;
 pub mod day08;
 pub mod day09;
 pub mod day10;
+pub mod day11;
