+++ 2023/advent_of_code/.supermaven/config.json
@@ -1,0 +1,1 @@
+{"link_id":"N6D0mMaT3rl4zYmHe2hwwToUmISyr9vO","machine_id":"IQFawFbsJE8ziqpgtWjbsRIgebiqa4lH"}
